"fix typo in `diff --git` for `/src/faceapi/routers/models.py.py` #4806. → 4781405↓“diff--git”"

+++ src/faceapi/routers/models.py
@@ -38,7 +38,7 @@ class PromptResponse(BaseResponse):
     height: Optional[int] = None
     negative_prompt: Optional[str] = None
     seed: Optional[int] = None
-    strength: Optional[int] = None
+    strength: Optional[float] = None
 
 
 class GeneratedReponse(BaseResponse):
